refactor: 향수 데이터 수집 로직 분리
