+++ acm-batch/src/main/kotlin/mashup/backend/spring/acm/collector/perfume/detail/PerfumeAccordVo.kt
@@ -1,0 +1,9 @@
+package mashup.backend.spring.acm.collector.perfume.detail
+
+data class PerfumeAccordVo(
+    val name: String,
+    val textColor: String?,
+    val backgroundColor: String?,
+    val opacity: Double?,
+    val width: Double?,
+)
